# Adding parallelism to audit proof generation