--- a/akd/src/append_only_zks.rs
+++ b/akd/src/append_only_zks.rs
@@ -740,6 +740,7 @@ impl Azks {
     }
 
     #[async_recursion]
+    #[allow(clippy::type_complexity)]
     async fn get_append_only_proof_helper<S: Database + 'static>(
         latest_epoch: u64,
         storage: &StorageManager<S>,
@@ -783,28 +784,26 @@ impl Azks {
                     if parallel_levels.map(|p| p as u64 > level).unwrap_or(false) {
                         // we can parallelise further!
                         let storage_clone = storage.clone();
-                        #[allow(clippy::type_complexity)]
-                        let tsk: tokio::task::JoinHandle<
-                            Result<_, AkdError>,
-                        > = tokio::spawn(async move {
-                            let my_storage = storage_clone;
-                            let child_node = TreeNode::get_from_storage(
-                                &my_storage,
-                                &NodeKey(left_child),
-                                latest_epoch,
-                            )
-                            .await?;
-                            Self::get_append_only_proof_helper::<_>(
-                                latest_epoch,
-                                &my_storage,
-                                child_node,
-                                start_epoch,
-                                end_epoch,
-                                level + 1,
-                                parallel_levels,
-                            )
-                            .await
-                        });
+                        let tsk: tokio::task::JoinHandle<Result<_, AkdError>> =
+                            tokio::spawn(async move {
+                                let my_storage = storage_clone;
+                                let child_node = TreeNode::get_from_storage(
+                                    &my_storage,
+                                    &NodeKey(left_child),
+                                    latest_epoch,
+                                )
+                                .await?;
+                                Self::get_append_only_proof_helper::<_>(
+                                    latest_epoch,
+                                    &my_storage,
+                                    child_node,
+                                    start_epoch,
+                                    end_epoch,
+                                    level + 1,
+                                    parallel_levels,
+                                )
+                                .await
+                            });
 
                         Some(tsk)
                     } else {