@slawlor 2023-01-27
New clippy lints required a LOT of changes... so that's the second commit lol